use a meshloader to load the wireRestShape instead of a component-dependant loader + draw rest shape
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/beamadapter/BeamAdapter@77 f2fb72c6-44b1-464a-b922-67ecee8ddb02

+++ SutureController.inl
@@ -260,7 +260,7 @@ bool SutureController<DataTypes>::wireIsAlreadyInitialized()
 
 	if (m_topology != NULL)
 	{
-		if (m_topology->getNbPoints() != numDofs || m_topology->getNbEdges() != (numDofs+numRigidPts-1))
+		if ((unsigned int)m_topology->getNbPoints() != numDofs || (unsigned int)m_topology->getNbEdges() != (numDofs+numRigidPts-1))
 			return false;
 	}
 	else
